+++ config/routes.js
@@ -1,7 +1,7 @@
 module.exports = {
   check: {
     h1: 'What happens when you refer yourself',
-    path: '/check',
+    path: '/what-happens-when-you-refer-yourself',
     title: 'What happens next',
   },
   results: {
@@ -9,11 +9,11 @@ module.exports = {
   },
   search: {
     h1: 'We need to know where your GP is',
-    path: '/search',
+    path: '/find-your-gp',
     title: 'Search for your GP',
   },
   start: {
-    h1: 'Find help with stress, anxiety or depression (England only)',
+    h1: 'Find a psychological therapies service (England only)',
     path: '/',
     title: 'Introduction',
   },