+++ test/integration/startPage.js
@@ -3,6 +3,7 @@ const chaiHttp = require('chai-http');
 const cheerio = require('cheerio');
 
 const constants = require('../../app/lib/constants');
+const iExpect = require('../lib/expectations');
 const routes = require('../../config/routes');
 const server = require('../../server');
 
@@ -14,22 +15,21 @@ describe('Start page', () => {
   let $;
 
   before('request page', async () => {
-    const res = await chai.request(server).get(`${constants.siteRoot}`);
+    const res = await chai.request(server).get(`${constants.siteRoot}${routes.start.path}`);
 
     $ = cheerio.load(res.text);
   });
 
   it('has a link to the next page', () => {
-    expect($('.button__start').text()).to.equal('Find help');
-    expect($('.button__start').prop('href')).to.equal(`${constants.siteRoot}${routes.check.path}`);
+    expect($('.nhsuk-button').text()).to.contain('Start');
+    expect($('.nhsuk-button').prop('href')).to.equal(`${constants.siteRoot}${routes.check.path}`);
   });
 
   it('has a link to the GP finder', () => {
-    expect($('.nhsuk-page-content').find('a').prop('href')).to.equal('https://www.nhs.uk/Service-Search/GP/LocationSearch/4');
+    expect($('main').find('a').prop('href')).to.equal('https://www.nhs.uk/Service-Search/GP/LocationSearch/4');
   });
 
-  it('has an urgent help call out', () => {
-    expect($('.samaritans__call').text()).to.equal('116 123');
-    expect($('.samaritans__email').text()).to.equal('jo@samaritans.org');
+  it('the breadcrumbs should have correct levels of links', () => {
+    iExpect.breadcrumbContent($, `${constants.siteRoot}${routes.start.path}`);
   });
 });