+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.23.0 / 2019-11-18
+===================
+- Use `nhsuk-frontend` for views, scss and JS
+- Remove redundant JS and SCSS replaced by `nhsuk-frontend`
+- Update site root to include `/service-search`
+
 0.22.0 / 2019-10-31
 ===================
 - Remove Greenkeeper - it is done automatically by the service